requirements.txt
asgiref==3.8.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
drf-yasg==1.21.10
easyocr==1.7.2
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.5.0
imageio==2.37.0
imutils==0.5.4
inflection==0.5.1
Jinja2==3.1.6
kiwisolver==1.4.8
lazy_loader==0.4
Levenshtein==0.27.1
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
mysqlclient==2.2.7
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.6
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pyclipper==1.3.0.post6
PyMySQL==1.1.0
pyparsing==3.2.3
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-Levenshtein==0.27.1
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
scikit-image==0.25.2
scipy==1.15.3
setuptools==80.8.0
shapely==2.1.1
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
tifffile==2025.5.10
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
